CKS-241 Add topics list page tests
