+++ functional-tests/support/action/windowIsScrolledToBottom.js
@@ -1,0 +1,5 @@
+module.exports = (selector) => {
+	return browser.execute(
+		() => window.innerHeight + window.pageYOffset >= document.body.offsetHeight
+	).value;
+};
